app/requirements.txt
Django==4.2.13
django-debug-toolbar==4.4.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.7
asgiref==3.8.1
crispy-bootstrap4==2024.1
django-crispy-forms==2.2
djangorestframework==3.15.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
django-htmx==1.18.0
